requirements.txt
# HTTP and HTML parsing
requests>=2.31.0,<3.0
beautifulsoup4>=4.12.0,<5.0
lxml>=4.9.0,<6.0

# PDF parsing
pdfplumber>=0.10.0,<1.0

# Entity resolution / fuzzy matching
rapidfuzz>=3.5.0,<4.0

# Data processing
pandas>=2.1.0,<3.0

# CLI
click>=8.1.0,<9.0

# Utilities
python-dateutil>=2.8.0,<3.0
pyyaml>=6.0,<7.0

# Dashboard
streamlit>=1.32,<2.0
plotly>=5.0,<6.0

# Testing
pytest>=7.4.0,<9.0
pytest-cov>=4.1.0,<6.0
